+++ src/main/java/kitchenpos/menus/tobe/domain/common/CleanName.java
@@ -2,7 +2,7 @@ package kitchenpos.menus.tobe.domain.common;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import kitchenpos.menus.tobe.domain.MenuNameValidationService;
+import kitchenpos.menus.tobe.domain.service.MenuNameValidationService;
 
 import java.util.Objects;
