+++ src/main/java/kitchenpos/menus/tobe/domain/entity/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.entity;
 
 import jakarta.persistence.*;
 import kitchenpos.menus.tobe.domain.common.*;
