+++ src/main/java/kitchenpos/menus/tobe/domain/repository/MenuRepository.java
@@ -1,0 +1,20 @@
+package kitchenpos.menus.tobe.domain.repository;
+
+import kitchenpos.menus.tobe.domain.entity.Menu;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface MenuRepository {
+
+    Menu save(Menu menu);
+
+    Optional<Menu> findById(UUID id);
+
+    List<Menu> findAll();
+
+    List<Menu> findAllByIdIn(List<UUID> ids);
+
+    List<Menu> findAllByProductId(UUID productId);
+}
