+++ src/test/java/kitchenpos/menus/domain/CleanNameTest.java
@@ -2,7 +2,7 @@ package kitchenpos.menus.domain;
 
 import kitchenpos.menus.application.FakeMenuPurgomalumClient;
 import kitchenpos.menus.tobe.domain.common.CleanName;
-import kitchenpos.menus.tobe.domain.MenuNameValidationService;
+import kitchenpos.menus.tobe.domain.service.MenuNameValidationService;
 import kitchenpos.products.infra.PurgomalumClient;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
